Move modernizr and jquery back to parent theme
I don't think that child themes should be responsible for including jquery and modernizr. They already exist in the theme as a parent, they should be loaded from the parent.
Keeping this as-is causes child themes to break as soon as they are created because the child theme will not include these files. Considering they will not be altered if copied to the child theme, I think they should stay with the parent.