Move sidebar width to pages
Moving the sidebar width out of the sidebar file and into the pages
makes it easier to customise the width of the sidebar according to the
width of the main content.

I can understand that having the columns class on the sidebar.php makes
it easier to change the size of the sidebar as a one off, but because
each of the main-contents would need to be edited anyway it makes more
sense to have them on the same page.