remove joints_remove_wp_ver_css_js() filter
Remove filter joints_remove_wp_ver_css_js() to allow CSS and JavaScript
to have ver paramters on the URLs, which allows for cache busting by
updating the version numbers.

Add code to use current theme version, as defined in style.css, as the
ver parameter on CSS and JavaScript loaded by the JointsWP theme in
place of the default ver parameter, which is the current version of
WordPress.

Fixes #49