Stop "gulp watch" from crashing on an undefined variable
If you're using "gulp watch", and happen to type a variable that's undefined, it will stop with an error about an unhandled error. The fix is to  pass gulp-plumber an error callback (Solution found here: https://github.com/yeoman/generator-gulp-webapp/issues/154#issuecomment-45990783).

Now instead of stopping, it will provide the error text in red, in the commandline.